games-emulation/dolphin: Do not remove mgba in 9999
Thanks to xarblu for the suggestion.

Closes: https://bugs.gentoo.org/846578

+++ games-emulation/dolphin/dolphin-5.0_p20220520.ebuild
@@ -117,7 +117,7 @@ declare -A KEEP_BUNDLED=(
 )
 
 src_prepare() {
-	if use mgba; then
+	if use mgba && [[ ${PV} != *9999 ]]; then
 		rmdir Externals/mGBA/mgba || die
 		mv "${WORKDIR}/mgba-${MGBA_COMMIT}" Externals/mGBA/mgba || die
 	fi
